+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 * Add `sync_standby` as a valid replica type for `cluster_has_replica`. (contributed by @mattpoel)
+* Add info and options (`--sync-warning` and `--sync-critical`) about sync replica to `cluster_has_replica`.
 * Add a new service `cluster_has_scheduled_action` to warn of any scheduled switchover or restart.
 * Add options to `node_is_replica` to check specifically for a synchronous (`--is-sync`) or asynchronous node (`--is-async`).
 * Add `standby-leader` as a valid leader type for `cluster_has_leader`.